fix(form): use singleton context pattern for SiblingImageInsertion
Move SiblingImageInsertionContext to _singletons/context/ to comply
with the project's lint rules for React context usage.
